A file that modifies how progress is checked.
Instead of simply tracking the number of intervals eliminated
it will instead track the log of the area that has not yet been
eliminated. Further work will need to be done to integrate this
code with the pre-existing IntervalChecks code to accurately
reflect the process of the solver. Also will need to have built
in tolerances to decide when the solver is finished solving.